Adicionando a alteração do pix conforme o valor do carrinho

+++ CookieByte/CookieByte/Model/OrderModel.swift
@@ -11,7 +11,7 @@ struct OrderModel {
     let user: String?
     let cookie: String
     let date: Date
-    let price: Int
+    let price: Double
     var qnt: Int = 1
     let pic: UIImage?
     let status: Bool
@@ -19,16 +19,12 @@ struct OrderModel {
 }
 
 class Order {
-    
     var cart = CartCard()
     
     static var shared = Order()
     
     var orders: [OrderModel] = [
-        OrderModel(user: "Gabriel", cookie: "Cookie Chocolate", date: Date(), price: 4, qnt: 1, pic: UIImage(named: "CookieT"), status: true, color: UIColor(named: "Cookie1Back")),
-//        OrderModel(user: "Gabriel", cookie: "Cookie Chocolate", date: Date(), price: 4.0, qnt: 1, pic: UIImage(named: "CookieM"), status: true, color: UIColor(named: "Cookie2Back")),
-//        OrderModel(user: "Gabriel", cookie: "Cookie Chocolate", date: Date(), price: 4.0, qnt: 1, pic: UIImage(named: "CookieB"), status: true, color: UIColor(named: "Cookie3Back")),
-//        OrderModel(user: "Gabriel", cookie: "Cookie Chocolate", date: Date(), price: 4.0, qnt: 1, pic: UIImage(named: "CookieT"), status: true, color: UIColor(named: "Cookie1Back")),
+        OrderModel(user: "Gabriel", cookie: "Cookie Chocolate", date: Date(), price: 4.0, qnt: 1, pic: UIImage(named: "CookieT"), status: true, color: UIColor(named: "Cookie1Back")),
     ]
     
     init() {}
